AdminSettings extracted and imported into Settings
