+++ src/screens/Settings/SettingsScreen.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import { ScrollView, Text, View } from 'react-native';
-import { useAppTheme } from '../../contexts/ThemeContext';
+import { ScrollView } from 'react-native';
 import { RootStackScreenProps } from '../../navigators/types';
-import ThemedClickableCardButton from '../../themedComponents/ThemedClickableCardButton';
+import AdminSettings from './AdminSettings';
+import UserSettings from './UserSettings';
 
 type Props = RootStackScreenProps<'Settings'>;
 
@@ -13,55 +13,6 @@ export type AdminSettingsSubComponentProps = {
 export type UserSettingsSubComponentProps = {
   navigation: Props['navigation'];
 };
-
-function AdminSettings({ navigation }: AdminSettingsSubComponentProps) {
-  const theme = useAppTheme();
-  return (
-    <View>
-      <Text>Admin-settings</Text>
-      <ThemedClickableCardButton
-        hideTitle
-        title="ProfileSettings"
-        content="Profile Settings"
-        iconName="user-cog"
-        iconColor={theme.colors.text}
-        onPress={() => navigation.navigate('Statistics', { period: 'today' })}
-      />
-      <ThemedClickableCardButton
-        hideTitle
-        title="HandleUsers"
-        content="Handle Users"
-        iconName="user-edit"
-        iconColor={theme.colors.text}
-        onPress={() => navigation.navigate('Statistics', { period: 'today' })}
-      />
-      <ThemedClickableCardButton
-        hideTitle
-        title="ShowAccessCode"
-        content="Show Access Code"
-        iconName="key"
-        iconColor={theme.colors.text}
-        onPress={() => navigation.navigate('Statistics', { period: 'today' })}
-      />
-      <ThemedClickableCardButton
-        hideTitle
-        title="ShowQRCode"
-        content="Show QR Code"
-        iconName="qrcode"
-        iconColor={theme.colors.text}
-        onPress={() => navigation.navigate('Statistics', { period: 'today' })}
-      />
-      <ThemedClickableCardButton
-        hideTitle
-        title="EditHousehold"
-        content="Edit Household"
-        iconName="home-edit"
-        iconColor={theme.colors.text}
-        onPress={() => navigation.navigate('Statistics', { period: 'today' })}
-      />
-    </View>
-  );
-}
 
 export default function SettingsScreen({ navigation }: Props) {
   const userIsAdmin = false; // You can toggle this for testing
